requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.6.0
pydantic==2.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1

